refactor: Move ImagePanResponder to <ImageArea />
